Removed unnecessary search index service
