--- ghost/core/core/server/services/search-index/index.js
@@ -1,8 +1,0 @@
-const SearchIndexService = require('./SearchIndexService');
-const PostsService = require('../posts/posts-service')();
-const models = require('../../models');
-
-module.exports = new SearchIndexService({
-    PostsService,
-    models
-});
